% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spicer-classify.R
\name{predict.spicer}
\alias{predict.spicer}
\title{Computes predictions from Spicer model.}
\usage{
\method{predict}{spicer}(model, kTest, type = "probability")
}
\arguments{
\item{kTest}{is of dimensions nTrain, nTest, length(model$sorted_kern_weight) (can pass extra kernels so long as ones referenced in model$sorted_kern_weight are included). In the case of multiclass prediction, the third dimension of kTest should include all possible kernels for each pairwise classification task.
type - applicable for classification (binary & multiclass) only - "response" returns the predicted class labels, while "probability" returns the probability of the positive class (the second class labels in model$opt$classes)}
}
\value{
Output is a prediction vector of length nTest computed by
f(x) = sum{model$sorted_kern_weigth[i]*KTest[,,i]}*model$comb_alpha + model$beta
if prediction task is 'regression', output is continuous values
if prediction task is 'classification' and type is 'response' - output are predicted labels
if prediction task is 'classification' and type is 'probability' - output is probabilities of positive class (model$opt$classes[2])
}
\description{
Computes predictions from Spicer model.
}
