% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spicer.R
\name{spicer}
\alias{spicer}
\title{spicer}
\usage{
spicer(K, yapp, C, opt = list())
}
\arguments{
\item{K}{N x N x M array. the (i,j,m)-element contains the (i,j)-element of the m-th kernel gram matrix.}

\item{yapp}{vector of length N with sample labels.  It should be a factor for binary/multiclass classification}

\item{C}{regularization parameter . Large values of C induce strong regularization. For L1 regularization C is a scalar; for elasticnet, C is a vector of length 2: C(1)|x| + C(2)x^2/2}

\item{opt}{list of options which control spicer behavior:
\describe{
 \item{loss}{type of loss function:'logit' (logistic regression, log(1+exp(- f(x)y))) for classification,
 'square' (square loss, 0.5*(y - f(x))^2) for regression}
 \item{regname}{type of regularization: 'l1' (default), 'elasticnet'}
 \item{optname}{optimization solver for dual variable (rho) inner minimization: 'Newton' (default), 'BFGS'}
 \item{stop_duality_gap}{TRUE/FALSE. If TRUE, Spicer employs duality gap for stopping criterion of outer loop. Default TRUE.}
 \item{stop_ineq_violation}{TRUE/FALSE. If TRUE, Spicer employs violation of inequality for stopping criterion of outer loop. Default FALSE.}
 \item{outer_maxiter}{maximum number of iterations of outer loop. (default 300)}
 \item{inner_maxiter}{maximum number of iterations of inner loop. (default 500)}
 \item{tol_outer}{tolerance of stopping criteria of outer loop. (default 0.001)}
 \item{tol_inner}{tolerance of stopping criteria of inner loop. (default tol_outer/1000)}
 \item{miniter}{minimum number of iterations of outer loop. (default 30)}
 \item{tol_miniter}{minimum relative improvement required for each outer iteration after miniter. (default 0.001)}
  \item{calpha}{increment factor of gamma: gamma^(t+1)=calpha*gamma^(t).  (default 10)}
  \item{incl_subw}{TRUE/FALSE. If TRUE, Spicer includes the full NxM matrix of alpha coefficients. Default FALSE.}
  \item{display}{1:display no progress messages, 2(default):display outer loop progress messages, 3:display inner loop progress messaages.}
  }}
}
\value{
A SPICER model with the following components:
  \describe{
  \item{comb_alpha}{N x 1 coefficient vector.}
  \item{kern_weight}{1 x M kernel weight vector, scaled to sum to 1}
  \item{bias}{bias term}
  \item{activeset}{indices of kernels that are active ({m : kern_weight[m] is not zero}).}
  \item{sorted_kern_weight}{vector of non-zero kernel weights sorted by magnitude, scaled to sum to 1.}
  \item{opt}{list of SPICER options used in run.}
  \item{history}{contains history of primal objective, dual objective, number of active kernels, and duality gap.}
  \item{kern_alpha}{If incl_subw is TRUE, the NxM matrix of alpha coefficients.}
  }
}
\description{
Port of Tomioka and Suzuki's SpicyMKL to R, expanded for multiclass and probability outputs.
}
\references{
\itemize{
  \item  V. Uzunangelov, C. K. Wong, and J. Stuart. Highly Accurate Cancer Phenotype Prediction with AKLIMATE, a Stacked Kernel Learner Integrating Multimodal Genomic Data and Pathway Knowledge. bioRxiv, July 2020.
  \item Suzuki,Tomioka.SpicyMKL: a fast algorithm for Multiple Kernel Learning with thousands of kernels. Mach Learn (2011) 85:77–108
  }
}
